Cleanup of new condtools for CTPPS
